requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pypdf
python-multipart
pydantic
streamlit
requests
httpx